Passage à 48 boutons de note :
Point technique clé : le masque de notes NotesMask passe de int (32 bits) à long (64 bits) — 48 boutons ne tiennent pas dans un int. Toutes les opérations de bits (1 << i → 1L << i, rising/falling/sustained/lastMask) sont élargies dans InputSnapshot, MappedControllerInput et PlayEngine.
STRING_COUNT = 48, ControllerMap : 48 slots. Défaut clavier = 4 rangées QWERTY grave→aigu (bas → rangée des chiffres, avec les touches OEM `, . / ; ' [ ] - = ``). Les 8 boutons F500 restent sur les 8 slots graves.
⚠️ Réalité importante à connaître : 48 notes distinctes ascendantes, ça marche pour les gammes (7 notes ≈ 7 octaves ≈ rentre dans la plage MIDI). Mais un accord de peu de notes (triade = 3 notes) voicé sur 48 boutons s'étale sur ~16 octaves → dépasse MIDI 127 ; le moteur wrappe (replie) les notes du haut au moment de jouer (donc des boutons hauts rejouent des notes plus basses). En clair : 48 boutons, c'est fait pour les gammes ; sur les accords les touches hautes bouclent. StringExpander empile serré puis décale la voicing pour garder le bas ≥ 0.

Tests : les assertions de plage strictes ([0,127], ≥ minPitch) ne tiennent plus à 48 pour les accords larges — remplacées par les invariants qui tiennent vraiment (compte = 48, toutes distinctes, ascendantes, plancher ≥ 0, classes correctes). 894 verts.

Migration config : ton config.json (21 slots) sera étendu à 48 par Normalize au prochain lancement (les slots ajoutés prennent les défauts). Pour le layout 48 propre, supprime %APPDATA%/VirtuoPhone/config.json.

Je n'ai pas relancé l'app (le headless prend la F500/l'audio). À tester avec F5 — surtout une gamme (pentatonique/mode) sur les 48 touches pour la pleine étendue. Changements non commités.

+++ Engine/PlayEngine.cs
@@ -123,7 +123,7 @@ public sealed class PlayEngine
 
         Direction dir = Direction.Neutral;
         GridCell current = grid[dir];
-        int lastMask = 0;
+        long lastMask = 0;   // 64-bit: one bit per note button (up to 48)
 
         int[] playingPitch = new int[liveButtons]; Array.Fill(playingPitch, int.MinValue);
         int[] streamId = new int[liveButtons];
@@ -151,7 +151,7 @@ public sealed class PlayEngine
 
         // Re-arm all voice state on the current instrument and re-trigger held buttons (used after Select /
         // audio rebuild). Local function so both paths share it.
-        void ReAttachInstrument(int notesMask)
+        void ReAttachInstrument(long notesMask)
         {
             liveButtons = Math.Min(instrument.GetStringCount(), input.NoteButtonCount);
             Array.Fill(playingPitch, int.MinValue);
@@ -160,7 +160,7 @@ public sealed class PlayEngine
             heldMono.Clear(); monoButton = -1; monoStreamId = 0;
             drone = instrument.GetDrone();
             drone?.OnPlayNoteUpdate(instrument.GetSoundPool(), current.Root);
-            for (int i = 0; i < liveButtons; i++) if ((notesMask & (1 << i)) != 0) heldMono.Add(i);
+            for (int i = 0; i < liveButtons; i++) if ((notesMask & (1L << i)) != 0) heldMono.Add(i);
 
             if (instrument.IsMonophonic())
             {
@@ -170,7 +170,7 @@ public sealed class PlayEngine
             else
             {
                 for (int i = 0; i < liveButtons; i++)
-                    if ((notesMask & (1 << i)) != 0)
+                    if ((notesMask & (1L << i)) != 0)
                     {
                         streamId[i] = instrument.Play(current.Voicing[i], i, 0f);
                         playingPitch[i] = basePitch[i] = current.Voicing[i];
@@ -214,8 +214,8 @@ public sealed class PlayEngine
             }
 
             InputSnapshot snap = input.Poll();
-            int rising = snap.NotesMask & ~lastMask;
-            int falling = lastMask & ~snap.NotesMask;
+            long rising = snap.NotesMask & ~lastMask;
+            long falling = lastMask & ~snap.NotesMask;
 
             // Select → next instrument.
             if (snap.Select && !lastSelect)
@@ -283,8 +283,8 @@ public sealed class PlayEngine
 
             if (instrument.IsMonophonic())
             {
-                for (int i = 0; i < liveButtons; i++) if ((falling & (1 << i)) != 0) heldMono.Remove(i);
-                for (int i = 0; i < liveButtons; i++) if ((rising & (1 << i)) != 0) { heldMono.Remove(i); heldMono.Add(i); }
+                for (int i = 0; i < liveButtons; i++) if ((falling & (1L << i)) != 0) heldMono.Remove(i);
+                for (int i = 0; i < liveButtons; i++) if ((rising & (1L << i)) != 0) { heldMono.Remove(i); heldMono.Add(i); }
 
                 int active = heldMono.Count > 0 ? heldMono[^1] : -1;
                 if (active != monoButton || (chordChanged && active >= 0))
@@ -299,9 +299,9 @@ public sealed class PlayEngine
                 if (chordChanged)
                 {
                     float gliss = instrument.GetGlissandoSeconds();
-                    int sustained = lastMask & snap.NotesMask;
+                    long sustained = lastMask & snap.NotesMask;
                     for (int i = 0; i < liveButtons; i++)
-                        if ((sustained & (1 << i)) != 0 && streamId[i] != 0)
+                        if ((sustained & (1L << i)) != 0 && streamId[i] != 0)
                         {
                             int newPitch = current.Voicing[i];
                             if (gliss > 0f)
@@ -322,7 +322,7 @@ public sealed class PlayEngine
                 }
 
                 for (int i = 0; i < liveButtons; i++)
-                    if ((rising & (1 << i)) != 0)
+                    if ((rising & (1L << i)) != 0)
                     {
                         streamId[i] = instrument.Play(current.Voicing[i], i, 0f);
                         playingPitch[i] = basePitch[i] = current.Voicing[i];
@@ -330,7 +330,7 @@ public sealed class PlayEngine
 
                 if (instrument.IsAutoLoop())
                     for (int i = 0; i < liveButtons; i++)
-                        if ((falling & (1 << i)) != 0 && streamId[i] != 0)
+                        if ((falling & (1L << i)) != 0 && streamId[i] != 0)
                         {
                             instrument.Stop(streamId[i]); streamId[i] = 0; playingPitch[i] = int.MinValue;
                         }
